api restructuration
